refactor: use less general function names
