+++ src/modules/timeDimensions.js
@@ -49,20 +49,6 @@ export const getTimeDimensions = () => ({
     },
 })
 
-export const getDefaulTimeDimensionsMetadata = () => {
-    return Object.values(getTimeDimensions()).reduce(
-        (acc, { id, name, dimensionType }) => {
-            acc[id] = {
-                id,
-                name,
-                dimensionType,
-            }
-            return acc
-        },
-        {}
-    )
-}
-
 export const getTimeDimensionName = (dimension, program, stage) => {
     if (!dimension.nameParentProperty || !program || !stage) {
         return dimension.name
